コードベースのリファクタリング
- useSwipe コンポーザブルを抽出し App.vue からスワイプロジックを分離
- useChecklistConfigSync コンポーザブルを抽出し Firestore/LocalStorage 同期ロジックを分離
- App.vue のデフォルトチェックリスト定数を useChecklistConfigSync に移動
- GenericChecklist.vue の不要な isActiveComputed ラッパーを除去
- firestore.ts の未使用関数 loadChecklistsConfig と loadChecklistData を削除
